+++ swayq.go
@@ -13,6 +13,7 @@ func main() {
 	var quietFlag = flag.Bool("q", false, "Do not print values")
 	var rawFlag = flag.Bool("r", true, "Print raw values")
 	var rawInputFlag = flag.Bool("R", true, "Read standard input as raw text lines")
+	var listFlag = flag.Bool("l", false, "List all available swayq modules")
 	var helpFlag = flag.Bool("h", false, "Show help")
 	flag.Parse()
 	var args = flag.Args()
@@ -25,11 +26,18 @@ func main() {
 
 	loader := swayqModuleLoader{}
 
-	if len(args) == 0 {
+	if len(args) == 0 || *listFlag {
 		modules, err := listModules()
 		if err != nil {
 			log.Fatalln(err)
 			os.Exit(1)
+		}
+
+		if *listFlag {
+			for _, module := range modules {
+				fmt.Println(module.name)
+			}
+			os.Exit(0)
 		}
 
 		fmt.Println("Available modules:")