+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 -   There is no longer a default module loaded. Instead, running `swayq` 
     without arguments will produce an overview of available modules. For 
     now, only non-builtin modules are shown.
+-   The `-l` flag lists available modules, separated by newlines.
 -   Positional arguments are now available in the `$ARGS` variable, in 
     the same format as stock `jq`.
 -   **Breaking:** You must now always provide a module. Running `swayq`